requirement.txt
torch
torchvision
numpy
pandas
tqdm
lmdb
opencv-python
pillow
tensorboardx
wandb


scipy
scikit-image
mpi4py 
blobfile
kornia


cv2
timm
natsort
openpyxl